update in note.txt

+++ app/Http/Controllers/Organizer/RoomController.php
@@ -68,7 +68,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error storing room: '.$th->getMessage());
+            logger()->error('Error storing room: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to store room',
@@ -164,7 +164,7 @@ class RoomController extends Controller
 
     public function update(Room $room, UpdateRoomRequest $request)
     {
-        $validated             = $request->safe()->only(['name', 'max_occupancy', 'description', 'price']);
+        $validated             = $request->safe()->only(['name', 'max_occupancy', 'description', 'price', 'is_active']);
         $validated['hotel_id'] = $request->user()->hotel->id;
 
         DB::beginTransaction();
@@ -191,7 +191,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error updating room: '.$th->getMessage());
+            logger()->error('Error updating room: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to update room',
@@ -209,13 +209,20 @@ class RoomController extends Controller
     {
         DB::beginTransaction();
         try {
+            if ($room->roomReservations->isNotEmpty()) {
+                return back()->with('alert', [
+                    'message' => 'You can\'t delete this room because it has reservations. Please change the status to inactive to disable this room.',
+                    'type'    => 'error',
+                ]);
+            }
+
             $this->deleteFile($room->cover_image);
             $room->delete();
 
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error deleting room: '.$th->getMessage());
+            logger()->error('Error deleting room: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to delete room',
@@ -260,7 +267,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error updating allotment: '.$th->getMessage());
+            logger()->error('Error updating allotment: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to update allotment',
@@ -311,7 +318,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error updating batch allotment: '.$th->getMessage());
+            logger()->error('Error updating batch allotment: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to update batch allotment',
@@ -355,7 +362,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error storing photo: '.$th->getMessage());
+            logger()->error('Error storing photo: ' . $th->getMessage());
 
             return response()->json([
                 'message' => 'Failed to store photo',
@@ -379,7 +386,7 @@ class RoomController extends Controller
             DB::commit();
         } catch (\Throwable $th) {
             DB::rollBack();
-            logger()->error('Error deleting photo: '.$th->getMessage());
+            logger()->error('Error deleting photo: ' . $th->getMessage());
 
             return back()->with('alert', [
                 'message' => 'Failed to delete photo',
